Refactor formatting functions to own module
This change is necessary to reduce the length of client.py in preparation
for adding examples to docstrings. Otherwise it would become unmanageably
long after examples are added.
